Revert "Feat: Stateful SDK [FR-129, FR-150, FR-167, FR-151]"
