+++ src/getModule/formatOutputs/utils.ts
@@ -1,52 +1,17 @@
 import { formatUnits, hexToString } from 'viem'
-import { TupleFormattedAbiParameter, Extras, TagCallback } from '../../types'
+import { TupleFormattedAbiParameter, Extras } from '../../types'
 import format from './format'
 
 type TupleCaseParams = {
   output: TupleFormattedAbiParameter
   res: any
   extras?: Extras
-  tagCallback: TagCallback
-}
-
-// The case for tuple outputs
-export const tuple = async ({
-  output,
-  res,
-  extras,
-  tagCallback,
-}: TupleCaseParams) => {
-  const formattedTuple: any = {}
-
-  await Promise.all(
-    output.components.map(async (c, index) => {
-      // try the name of the component, if it doesn't exist, use the index
-      formattedTuple[c.name ?? `_${index}`] = format({
-        output: c,
-        res: res[c.name ?? index],
-        extras,
-        tagCallback,
-      })
-    })
-  )
-
-  return formattedTuple
-}
-
-// The case for tuple[] outputs
-export const tupleArray = ({ res, ...rest }: TupleCaseParams) =>
-  res.map((resI: any) => tuple({ res: resI, ...rest }))
-
-// The case for decimal outputs
-export const formatDecimals = (value: bigint, decimals?: number) => {
-  if (!decimals) throw new Error('No decimals provided')
-  return formatUnits(value, decimals)
 }
 
 // Tag: "any" tries to parse the result as JSON from hex, if it fails-
 // tries to parse the result as a string from hex, if it fails-
 // returns "Data is not a valid JSON string"
-export const formatAny = (res: any) => {
+export const any = (res: any) => {
   try {
     return JSON.parse(hexToString(res))
   } catch {
@@ -56,4 +21,30 @@ export const formatAny = (res: any) => {
       return 'Data is not a valid JSON string'
     }
   }
+}
+
+// The case for tuple outputs
+export const tuple = ({ output, res, extras }: TupleCaseParams) => {
+  const formattedTuple: any = {}
+
+  output.components.forEach((c, index) => {
+    // try the name of the component, if it doesn't exist, use the index
+    formattedTuple[c.name ?? `_${index}`] = format(
+      c,
+      res[c.name ?? index],
+      extras
+    )
+  })
+
+  return formattedTuple
+}
+
+// The case for tuple[] outputs
+export const tupleArray = ({ res, ...rest }: TupleCaseParams) =>
+  res.map((resI: any) => tuple({ res: resI, ...rest }))
+
+// The case for decimal outputs
+export const decimals = (value: bigint, extras?: Extras) => {
+  if (!extras?.decimals) throw new Error('No decimals provided')
+  return formatUnits(value, extras.decimals)
 }
