+++ src/types/index.ts
@@ -2,4 +2,3 @@ export * from './base'
 export * from './parameter'
 export * from './utils'
 export * from './guards'
-export * from './tag'
