--- src/utils/tagProcessor/approval.ts
@@ -1,3 +1,0 @@
-import getRequiredAllowance from '../getRequiredAllowance'
-
-export default getRequiredAllowance
