--- src/utils/tagProcessor/index.ts
@@ -1,7 +1,0 @@
-import approval from './approval'
-import decimals from './decimals'
-
-export default {
-  decimals,
-  approval,
-}
