+++ tests/getModule/getModule.test.ts
@@ -1,7 +1,7 @@
 import { expect, describe, it } from 'bun:test'
 
 import getModule from '../../src/getModule'
-import { getTestConnectors } from '../testHelpers/getTestConnectors'
+import { getTestConnectors } from '../getTestConnectors'
 import utils from '../../tools'
 
 describe('Get A Module', () => {
