+++ tests/getWorkflow/headlessWorkflow.test.ts
@@ -1,6 +1,6 @@
 import { expect, describe, it } from 'bun:test'
 
-import { getTestConnectors } from '../testHelpers/getTestConnectors'
+import { getTestConnectors } from '../getTestConnectors'
 import utils from '../../tools'
 import getWorkflow from '../../src/getWorkflow'
